+++ src/BroString.h
@@ -33,13 +33,13 @@ public:
 	typedef IdxVec::const_iterator IdxVecCIt;
 
 	// Constructors creating internal copies of the data passed in.
-	BroString(const u_char* str, int arg_n, int add_NUL);
+	BroString(const u_char* str, size_t arg_n, int add_NUL);
 	explicit BroString(const char* str);
 	explicit BroString(const std::string& str);
 	BroString(const BroString& bs);
 
 	// Constructor that takes owernship of the vector passed in.
-	BroString(int arg_final_NUL, byte_vec str, int arg_n);
+	BroString(int arg_final_NUL, byte_vec str, size_t arg_n);
 
 	BroString();
 	~BroString()	{ Reset(); }
@@ -49,19 +49,19 @@ public:
 	bool operator<(const BroString& bs) const;
 
 	byte_vec Bytes() const	{ return b; }
-	int Len() const	{ return n; }
+	size_t Len() const	{ return n; }
 
 	// Releases the string's current contents, if any, and
 	// adopts the byte vector of given length.  The string will
 	// manage the memory occupied by the string afterwards.
 	//
-	void Adopt(byte_vec bytes, int len);
+	void Adopt(byte_vec bytes, size_t len);
 
 	// Various flavors of methods that release the string's
 	// current contents, if any, and then set the string's
 	// contents to a copy of the string given by the arguments.
 	//
-	void Set(const u_char* str, int len, int add_NUL=1);
+	void Set(const u_char* str, size_t len, int add_NUL=1);
 	void Set(const char* str);
 	void Set(const std::string& str);
 	void Set(const BroString &str);
@@ -96,7 +96,7 @@ public:
 	//
 	// Note that you need to delete[] the resulting string.
 	//
-	char* Render(int format = EXPANDED_STRING, int* len = 0) const;
+	char* Render(int format = EXPANDED_STRING, size_t* len = 0) const;
 
 	// Similar to the above, but useful for output streams.
 	// Also more useful for debugging purposes since no deallocation
@@ -122,7 +122,7 @@ public:
 	// A negative @length means "until end of string".  Other invalid
 	// values result in a return value of 0.
 	//
-	BroString* GetSubstring(int start, int length) const;
+	BroString* GetSubstring(int start, size_t length) const;
 
 	// Returns the start index of s in this string, counting from 0.
 	// If s is not found, -1 is returned.
@@ -145,7 +145,7 @@ protected:
 	void Reset();
 
 	byte_vec b;
-	int n;
+	size_t n;
 	unsigned int final_NUL:1;	// whether we have added a final NUL
 	unsigned int use_free_to_delete:1;	// free() vs. operator delete
 };
